// src/app/accounts/models/api/account.ts
/**
 * OMRO DigitalApp API
 * API used on OMRO Digital Resources
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Account { 
    id?: string;
    iban?: string;
    name?: string;
    currencyCode?: string;
    workingBalance?: number;
    color?: string;
}